Fixed setup.py `__version__` relative path warning
When I try to run the `setup.py` file from a parent folder, like is the
case when the perception module is a submodule of a perent module, I get
the following error:

Traceback (most recent call last):
  File "./perception/setup.py", line 25, in <module>
      exec(open('perception/version.py').read())
      FileNotFoundError: [Errno 2] No such file or directory:
      'perception/version.py'

By making the path absolute this error is fixed. This seems to work with
the `python setup.py install` and `pip install` commands. I do however
not know if it works with all the setup methods.